Fix issue with setting empty values
Make most properties pointers in order to be able to distinguish between
null-values and unset-values (see go-github blog post discussing the
same problem for go-github [0]) -- it's actually the same problem described
in gambol99/go-marathon#124 but for all^W most properties.

I left out sub-properties such as health checks as I either pass in an
array of new health checks that will replace the current health-checks
or I omit health checks, which means that there should not be a change.

Instead of providing convenience functions such as
`github.String("string")` that return a pointer to a string, I added
methods that use the builder pattern'ish approach that quite a few
properties had before as well, i.e. construction looks something like:

    application := marathon.NewDockerApplication().
      Name("myapp").
      Command("/hello-world-server").
      AddEnv("PORT", "8080").
      ...
      Count(0)

    application.Container.Docker.Container("alpine-hello-world-server").
      Bridged().
      ExposePort(8080, 0, 0, "tcp").
      SetPrivileged(false).
      SetForcePullImage(false).
      Parameter("ulimit", fmt.Sprintf("nofile=%d", concern.Limits.FileMax))

There are specialized methods called for example `EmptyArgs()` to explicitly set
an array or hash to be empty (vs. omiting it, which would mean it keeps the current
value).

For consistency's sake, I renamed `Arg` -> 'AddArgs' as this change is a breaking
change anyways.

Note: this is a request for comments, I am happy to change/add stuff around or go
with a different approach -- I would just like to use the lib in project and need
a fix for this problem.

[0] https://willnorris.com/2014/05/go-rest-apis-and-pointers